# internal/swaggerdocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Private bank API — syncs Fio Bank transaction data into Postgres, matches member payments to their bank transactions, and detects missed dues. Dev-only documentation (see config.EnableSwaggerDocs): every route below requires a Keycloak bearer token unless noted otherwise.",
        "title": "Bank System API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/payments/waivers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Requires the manage-transactions role. All members, newest-created first.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "List every waiver",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.paymentWaiverResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/payments/{member_number}/waive": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Requires the manage-transactions role. Idempotent (repeat calls keep the original reason). 400 if the month is already covered by a real payment.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Write off a member's missed month",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Member number",
                        "name": "member_number",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Month to waive, and why",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.waiveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.paymentWaiverResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/payments/{member_number}/waive/{year}/{month}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Requires the manage-transactions role. The month reappears in missing-payment lists on the next read.",
                "tags": [
                    "payments"
                ],
                "summary": "Undo a waiver",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Member number",
                        "name": "member_number",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Waived year",
                        "name": "year",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Waived month, 1-12",
                        "name": "month",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "no content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.paymentWaiverResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "member_number": {
                    "type": "integer"
                },
                "month": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "handler.waiveRequest": {
            "type": "object",
            "properties": {
                "month": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Keycloak-issued bearer token for bank-system's Keycloak client, sent as `Authorization: Bearer \u003ctoken\u003e`. Which role(s) a given route requires is noted on that route below — see internal/keycloak.Role for the full list.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}